fix(l2_cleanup): remove redundant cast in byte accumulation
m.size is already u64; the previous as u64 cast was a no-op that
silenced a clippy lint without adding correctness.

+++ nodedb/src/event/collection_gc/l2_cleanup.rs
@@ -167,7 +167,7 @@ async fn delete_prefix(
     while let Some(meta) = list.next().await {
         match meta {
             Ok(m) => {
-                total_bytes += m.size as u64;
+                total_bytes += m.size;
                 if let Err(e) = store.delete(&m.location).await
                     && first_err.is_none()
                 {
